backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI orchestration service for the LLVM energy analyzer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.11.0",
    "pytest>=9.0.3",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.46.0",
]

[project.scripts]
backend = "backend:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"
